feat: [PIE-4272]: file and folders restructure
